+++ .isort.cfg
@@ -19,7 +19,6 @@ known_first_party=
     legacy,
     regulations3k,
     search,
-    transition_utilities,
     v1,
     wellbeing,
     permissions_viewer,
